+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Property keys in command help output are now sorted alphabetically
 - UTF-8 characters in output are now the default regardless of whether stdout is a TTY. Added --utf8 flag to explicitly set this behavior.
+- Show now toggle-disk status ("Adding Disk"/"Removing Disk") to the State column of "resource list" and "volume list"
 
 ### Fixed